Add data dir
Spacemacs uses ~/.emacs.d in 2 ways: to store the code implementing the
spacemacs framework and to track state. When we tell spacemacs that it lives
in the store it tries to use the store location to track state as well. This
patch adds a new variable, spacemacs-data-directory, for keeping track of state.
This defaults to spacemacs-start-directory for upstream compatibility.