Inhibit read only
The first time that a user runs spacemacs it will open the default
configuration from the source tree and modify the buffer. The buffer is
read-only since the file it opened is, but since we save the file to the users
home directory and not back to it's original location temporarily inhibiting
read-only solves our problem.